Migrando la vista del rank de escuelas a BS4 y typescript (#3782)
